pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bloomberg", "configs", "instruments", "services"]

[project]
name = "marsapilatam"
version = "0.1.0"
description = "Bloomberg MARS API client — Latin America"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "PyJWT>=2.8",
    "tenacity>=8.2",
    "pandas>=2.2",
    "numpy>=1.26",
    "python-dateutil>=2.9",
    "streamlit>=1.35",
    "plotly>=5.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py311"
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "ANN"]
ignore = ["ANN401"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
